Added the ability to add fishing spots from the GUI. Added the ability to remove fishing spots from the GUI. Automatic detection and management of modes: tactics, creature, and combat lock, once per script launch. Fixed Combat Timeline. Fixed hotkeys.
The script has a fishing mode and a combat mode; combat can be enabled, and the combat-only mode can also be enabled.

The bot now manages: Fishing, optimized fishing path to the next map, combat detection, and combat positioning during the combat phase (closest to the enemy).

The next update should handle fights automatically.
(The end-of-combat popup is already handled)

-Formatted Comments